Abstraction for architecture dependent secure memset.
In ecc_dh.c it is widely used to zero out the memory used for holding
sensitive information, such as private keys. This memory is cleared by
a memset, but as this is done right at function return, this memset
will most likely be optimized out by the compiler, as the memory will
no longer be used, hence no need to set it. To prevent the memset from
being optimized out, it is necessary to place a memory barrier at the
memset, but such a barrier is compiler depended.

To solve the above, a _set_secure function has been introduced. This
defaults to an inlined function calling memset and, in cases where the
compiler defines __GNUC__, adds a memory barrier which ensures that
the memset is not optimized out. For compilers, not defining the
__GNUC__, it may be necessary to define TINYCRYPT_ARCH_HAS_SET_SECURE
which then declares the _set_secure function extern, thus meaning that
it will require target implementation, which in turn allows for
architecture/compiler dependent implementation.

This also fixes the memset's in uECC_make_key and uECC_make_key_with_d
which actually was missing the memory barrier.

Fixes #32

Signed-off-by: Danny Oerndrup <daor@demant.com>